requirements.txt
asgiref==3.7.2
cachetools==5.5.1
certifi==2023.7.22
charset-normalizer==3.4.1
cloudinary==1.36.0
cssselect==1.2.0
cssutils==2.11.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-appconf==1.0.5
django-ckeditor-5==0.2.10
django-compressor==4.4
django-import-export==3.3.2
django-jazzmin==2.6.0
django-mathfilters==1.0.0
django-taggit==5.0.1
djangoql==0.17.1
dkimpy==1.1.8
dnspython==2.7.0
et-xmlfile==1.1.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
lxml==5.3.0
MarkupPy==1.14
more-itertools==10.6.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
ply==3.11
premailer==3.10.0
# psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
pywin32-ctypes==0.2.3
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.32.3
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
yagmail==0.15.293
